pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cbt-llm"
version = "0.0.1"
description = "CBT + SNOMED Knowledge Graph Pipeline"
authors = [
  { name = "Your Name" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

dependencies = [
    "neo4j",
    "python-dotenv",
    "pandas"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["cbt_llm"]
force-sort-within-sections = true